incorporate CR feedback
