--- components-rs/components.rs
@@ -1,6 +1,0 @@
-//! Facade for the common Rust components.
-
-#[cfg(not(standalone_profiler))]
-pub use crate::{
-    agent_info, bytes, ffe, log, remote_config, sidecar, stats, telemetry, trace_filter,
-};
