+++ ext/configuration.c
@@ -16,17 +16,26 @@ ZEND_EXTERN_MODULE_GLOBALS(datadog);
 
 #ifndef DDTRACE
 bool datadog_alter_dd_service(zval *old_value, zval *new_value, zend_string *new_str) {
-    UNUSED(old_value, new_value, new_str);
+    UNUSED(old_value, new_value);
+    if (DATADOG_G(request_initialized)) {
+        ddtrace_sidecar_submit_span_data_direct(&DATADOG_G(sidecar), NULL, new_str, get_DD_ENV(), get_DD_VERSION());
+    }
     return true;
 }
 
 bool datadog_alter_dd_env(zval *old_value, zval *new_value, zend_string *new_str) {
-    UNUSED(old_value, new_value, new_str);
+    UNUSED(old_value, new_value);
+    if (DATADOG_G(request_initialized)) {
+        ddtrace_sidecar_submit_span_data_direct(&DATADOG_G(sidecar), NULL, get_DD_SERVICE(), new_str, get_DD_VERSION());
+    }
     return true;
 }
 
 bool datadog_alter_dd_version(zval *old_value, zval *new_value, zend_string *new_str) {
-    UNUSED(old_value, new_value, new_str);
+    UNUSED(old_value, new_value);
+    if (DATADOG_G(request_initialized)) {
+        ddtrace_sidecar_submit_span_data_direct(&DATADOG_G(sidecar), NULL, get_DD_SERVICE(), get_DD_ENV(), new_str);
+    }
     return true;
 }
