+++ tracer/mod.rs
@@ -1,0 +1,3 @@
+//! Rust facade for tracer support.
+
+pub use crate::{ffe, stats, trace_filter};
